VIC-1647 vicos toolchain setup (#975)
Phase 1 of vicos (linux) toolchain integration

We are switching to a building with a new toolchain for embedded linux.
This change adds a script to fetch the toolchain and provides a preliminary
integration into the build system, so that vicos builds can live along side
our working linux-android system.

Until all build targets work with the new toolchain, we'll keep using
the 'android' target. Anyone who needs to work on the vicos toolchain
builds can build `vicos-staging`. The `vicos` is intended for any binaries
that can run as standalone linux executables and either can not or are not
also built for android.

Detailed summary:

- Create toolchain file for vicos toolchain

- Add a vicos target to build-victor

- vicos support for BLECozmo lib

- vicos support for das-client

- ankiutil changes to check for VICOS define

- Workaround for install_rpath issue with civetweb

- Add a vicos-staging platform entry

- Add vicos.py and update toolget.py to pull down vicos-sdk

- Remove libandroid dependency from util

- Remove util dep from cti-messaging

- Update audio submodule for vicos toolchain

* lib/audio ce4937a:
  > Merge pull request #93 from anki/vic-1647_vicos-toolchain-support
  > Merge pull request #92 from anki/vic-1692_linux-build-support