Vic 1692 linux build support (#973)
* Platform-specific scripts for fetching deps

Linux systems do not have homebrew, so we need a different way to determine
whether build deps are available and install them.  This change preserves the
mac methods and adds simple checks for linux + instructions that might unblock
building. We'll need a better solution, but this change should bootstrap the process
of building on linux.

* Get CMAKE_EXE path using cmake.py script

Using cmake.py, we can install/determine the path of our baked-in
cmake executable.  This replaces the mac-specific path that was previously used.

* Search for flatc host executable in cte prebuilts

* Update DEPS for flatc prebuilt support

* Only use unbuffered output in dependencies.py

* Explicitly name target_triple variable