VIC-1647 Fix URL for vicos toolchain on linux host (#997)
We currently use a prebuilt version of clang for ubuntu, so for now the
vicos-linux toolchain needs this distro.

Ensure that we get the path to the toolchain even if it is installed
from invoking vicos.py in the build script. Most people don't need the
vicos-sdk (toolchain) yet.  Until we are ready to transition, we need
to keep the vicos.py invocation out of the fetch-build-deps script
and ensure that the build script will install it as necessary for
whoever needs it.