VIC-1752 vic-ble-cli: support WEP, hidden, or open WiFis
This goes along with the change in the Victor OS' vicdevsetup
to allow the configuration of WEP, hidden, and open networks.
This change attempts to support older robots that do not have
the new change by sending the old wifi configuration command
along with the new one, but there are no guarantees.  I also
tried to update some of the existing scripts in this directory
that used wifi-set-config to use the new format.  I had no easy
means to test them however.