VIC-1747: Fix crash during vic-anim shutdown (#996)
* VIC-1747: Fix crash during vic-anim shutdown

Rework CozmoAnimMain to handle SIGTERM without crashing.
It's generally not safe to do stuff in a signal handler because it's not synchronized
with the main thread.

Fix here is to set a flag, then have main loop perform shutdown.

Also added checks for failed init and failed update.
In case of error, process exits with an appropriate status code.