Fix cloud process being unable to reconnect to engine (#1008)
If the cloud process crashes, it needs to be able to reconnect to
the servers it talks to in the engine. Because LocalUdpServer
currently tries to permanently bind its socket to the first client
that connects to it, subsequent connection attempts fail if the
cloud process crashes and tries to reconnect. This change adds an
option to stop the server from doing this, which means that
subsequent reconnect attempts from the cloud process will succeed.