VIC-1792 Keepaway - Prox For Distance (#1035)
Adding a basic implementation of the prox sensor for distance
measurement in Keepaway. If the target is visible, and is seen to be
within the FOV of the prox sensor, the prox measurement will be used
for distance to the target. This allows the behavior to be much more
responsive to very close cubes, but will still consider the cube to be
invisible the marker is obscured for >  targetVisibleTimeout_s. The
new prox functionality defaults to enabled and can be disabled at run
time via BehaviorInstancParams on the consolevars web service or the
default can be changed in Keepaway.json