VIC-632 Use HS serial port if available (#1075)
We now attempt to open + configure the high speed serial port (`/dev/ttyHS0`) by default.
If that fails and we are building on a robot platform (android), we attempt to use the
old `/dev/ttyHSL1` device.

Since old versions of the OS might contain a startup script that creates a file named
`/dev/ttyHS0` during boot, we need to retry after attempting to open and configure
the serial port (e.g. calling tcgetattr). Failure to check this case broke the previous
attempt to implement this functionality (b605d490).