VIC-1436 ReactToMotion behavior (#1038)
A behavior to look at motion, and if it's persistent, turn toward it.
The behavior holds a list of ConditionMotionDetected, each looking at
a different region (left, right, up), which are used both to determine
if the behavior should be activated, and to control it once active.
When a condition is true, the eyes will turn in the specified
direction. Then, if no motion occurs for a while, the behavior ends.
Or, if no motion in the same position occurs for a (shorter) while, but
motion occurs elsewhere, the eyes move to the new position. If motion
is recurring in an area for some time, the robot turns toward it. It
then starts looking for motion again, and repeats the whole process if
some is found, or otherwise exits.
There's a version of eyes and turns using animations, and one using
the procedural eyes and turn actions.
Also, the MotionDetected conditions now return false and turn off the
vision mode if the robot moves under its own accord or is picked up.
Also fixes unit tests in xcode.