Remove CameraService reference in robot.cpp::SyncTime (#1125)
This ensures that we are not capturing frames unnecessarily.

In `SendSyncTime`, we were obtaining a singleton instead of `CameraService`
in order to obtain a timestamp.  This was causing the CameraService to be initialized,
which starts capturing camera frames.  This change replaces the singleton reference
with the inlined code to obtain the timestamp.