wifi: add an agent to handle hidden AP
Instead of having to fiddle with the service string to convert an
unknown hidden service to a known one, it's possible to register
an agent for connman via D-Bus, so that a callback is called to
get the info necessary to connect to the hidden AP specified.

This patch implements such agent for the WPA[2]-PSK case.