VIC-1952 build-victor: get correct path to cmake
If we haven't downloaded and installed cmake, then cmake.py will
download it and print status messages to stdout before finally
printing the path to cmake as the last line of output.  Previously, in
build-victor.sh, we captured all of these many lines of output and
assigned it to CMAKE_EXE.  This caused a later invocation of CMAKE_EXE
to fail since it was not the path to cmake.  With this change, cmake.py
will still output verbose messages during the installation of cmake, but
it will also support a separate command, --find-cmake, that will just
try to find cmake at the required version and output its path.
build-victor.sh and envsetup.sh are modified to use both.

(cherry picked from commit de5c8a817f8692b12fec8e0311ba2e361745fd0e)