VIC-2129 Allow animations to complete while audio is still playing
Previously on Cozmo the animations had to wait for audio to
complete to finish the animation, however, in we don’t have that
constraint on Victor. If an audio event plays over the end of an
animation it will simply continue to play out. The Audio Designers
can handle the edge cases where the next animation audio events
begin to play while the previous animation’s audio is still playing.

- Update Animation Streamer End of animation
- Add Init Animation method to AnimationAudioClient