VIC-1952 Call install.sh from deploy.sh
When we use bitbake to build the Victor apps/libs we will call
install.sh to install the apps/libs in the appropriate place to be
included with the O/S image.  We want to have 1 script that does that
whether we are building an O/S image or just doing a live deploy to
a robot.  This change separates the install portion of deploy.sh into
a separate script.

(cherry picked from commit 661c0c619232250471d4ed75f4db021e6c97af30)