Warn when HLAI behaviors dont want to activate (#1297)
If a state transition brings HLAI from state A to state B, we don't
check if the behavior of state B wants to be activated. This is a bit
weird given how the rest of the behavior system works, but it does
allow for transitions from A to C in two ticks, by going through B,
even if the behavior in B doesn't want to activate.
This adds a warning if the behavior in B doesn't activate upon state
transition, and a second, more important warning if there is no
transition to C in the next tick.