VIC-2188 victor_deploy - Delete stale assets
After updates or switching branches, it is possible that the contents of
the '/anki' directory on Victor contain stale assets.  Loading and using
these stale assets could cause the Victor apps to crash or exhibit
undefined behavior.

Previously, when we deployed assets to Victor, we didn't delete files on
Victor that were not present in our staging directory.  This left the
stale assets in place.

To fix this, I had to further separate out the phases of deployment and
have victor_deploy.sh first "stage" the build artifacts from both the
android build and the vicos build into the same place.  From there,
deploy.sh can be called 1 time and told to delete stale assets on
Victor.