VIC-2334 Modify deploy.sh to remount rootfs
If the rootfs is mounted read-only, we need to temporarily remount it as rw in order to deploy
into /anki. This implementation adds a verbose log to make it easier to follow steps (if desired).
We then run the actual deploy actions in a subshell and capture the result so that we
can attempt to remount the rootfs back to ro in case deploy fails for some reason.