Compound BEIConditions (#1316)
We might regret this later, but this opens the flood gates for compound
bei conditions like "not on charger and (stimulation is high or
a timer has expired)". Instead of composing conditions from an
AndCondition, OrCondition, and Negate, this creates one single Compound
condition that manages and evaluates all conditions. This should
hopefully ease debugging, since you won't have to step in and out of
multiple conditions during their evaluation.
Previous uses of the Negate condition are moved over to Compound.