VIC-2406 Use rsync daemon for syncing files (#1350)
@alchaussee and @kevinatanki have reported crashes/hangs in sshd during rsync deploys.
This change reverts back to using rsync in daemon mode for file transfers.
This turns out to be more stable and slightly faster than running rsync over ssh, at the
expense of extra complexity to install/start/stop the daemon. This time around, we manage
rsyncd with a systemd service instead of starting it inside the deploy script. This makes
starting/stopping rsyncd more robust.  We may be able to eventually move to installing
rsync 2.6.9 in the developer OS image in the future.

Note that this change also creates a /anki-devtools dir on the robot and installs rsync.bin
into /anki-devtools/bin.  This allows us to easily keep track of the fact that rsync is
intended to be a developer tool and should never ship.  The rsyncd.{conf,service} files
are deployed to /run which is a tmpfs.

(cherry picked from commit 6bb0b477b1a508cfee906816374be640e72d0866)