VIC-1850: Strawman proposal for DAS logging format (#1294)
* VIC-1850: Strawman proposal for DAS logging format

Minimal implementation to demonstrate how log events will be captured, formatted, and passed into android log buffers.

This version is compatible with existing log calls.  Any message that passes service-level filtering will be garnished with DAS attributes,
formatted into a json-like syntax, then submitted to the android log system.

Events are written to the "radio" buffer which appears to be unused on victor. As is, log events are sent once to the default buffer as "plain text", then again as an encoded st truct.
We could eliminate the "plain text" version, but developers would then need a way to view the encoded version.

Log consumers (TBD) will read incoming events from the android log system, parse and filter as needed, then uploaded to DAS analytics or splunk endpoint.

I have not yet addressed requirements for DAS v2, but they will be incremental. Jason L is working on a macro-style declaration that can be used as an alternative to the existing Log::sEvent() method.

Put DAS helper classes into new library to be shared between services.