NavMap static HandlesType instead of Concrete (#1322)
Navmap data objects have non-default constructors that take the
EContentType and perhaps more. In a future PR, a single data class
will handle multiple EContentTypes (e.g., a MemoryMapData_Obstacle
could handle prox obstacles, cliffs, etc.). Because of this, the
protected default constructor won't be able to tell us if a data
class could be used for a given EContentType.
This replaces the sneaky Concrete helper, which default-constructs
a data type to determine its EContentType, with a static call to
HandlesType.