VIC-1672 Cube Detection Prototype (#1388)
Adds a new behavior called "InspectCube" for evaluating the
effectiveness of triggering keepaway using natural cube based
interaction like "teasing" a cat with a toy. Currently this behavior
can only be set from webViz for evaluation. It will need to be tied
likely be tied to a higher level component to be triggered from some
external input, possibly as a state in HLAI accessed if we think the
user is holding the cube and doing something with it.