Fix leak in engine
The leak was caused by sections of code that are specifically not executed after the robot has been packed out. One of these sections of code was AIComponent::Update() which calls BehaviorComponent::Update() which is responsible for clearing a vector. That vector is populated by a RobotToEngine message handler which I assume was handling something like robot state or some other fairly frequent message. The vector would grow forever as it was not being cleared by the Update() and eventually we would run out of memory.

The fix is to just not subscribe to the messages once the robot has been packed out so nothing is ever added to the vector. This won't cause any problems as no behaviors are running after pack out so it doesn't matter if they can't subscribe to messages.