VIC-2328 Render noise outside of the face renderer (#1453)
* VIC-2328 Render noise outside of the face renderer

The face rendering pipeline consists of stages, each depending on the previous stage,
e.g. drawing each eye, transforming the face, adding scanlines, distortion, noise,
with the final image converting to a RGB565 texture to be sent to the robot.

Each state in the pipeline follows a similar path:
1) Did an earlier stage result in nothing to draw?
     If yes, early out. Frequently this is a transform that results in zero area
     eyes or face
2) Did an earlier stage change the input to this stage?
     If yes, then this stage must render
     If no, then compare the latest face data with the face data used to generate
     the cached image, if the face data has changed this stage must render
3) Render this stage
     Update the cached face data with the latest face data
     Use the face cache of the previous stage as input
     Assign a face cache for this stage as output
     Apply the stage
4) Don't render this stage
     Leave the cached face data as it was
     Take the input from the previous stage and pass it on as the output from this
     stage, no image copy required
5) Keep track of the latest face cache, this will be the final image

A key concept here is passing indexes into the cache for inputs and outputs, e.g.
if there is no face transform then the input to the noise stage can be the output
from the eyes stage. This is an alternative to the face transform performing a memcpy
from it's input to it's output. However, it comes at a cost of making all surfaces
RGB instead of keeping some I8. It's possible that color can be applied after the
noise has been added, potentially during the HSV conversion.

The code for each stage is kept largely as the original.
Scanline distortion has changed from modifying the input image to generating a new
output image, this allows eye rendering and face transforms to be cached and the
scanline distortion applied to the original output rather than a face that has already
had scanline distortion applied.
Similarly for the noise.

ApplyScanlines is a special case as it is part of the public API and used elsewhere,
it's functionality has been retained and does not affect the face cache.

Removes previous option of rendering if eye+face parameters have changed,
this is now part of the pipeline

Fixes references to an empty glow surface when glow is disabled

Fix brackets to match the rest of the file

Hoisted some asserts from inner loops to their outer scope

* PR review

* PR review