CLAI-160 Create + use CLAD emitter for Go (#1492)
* Step 1: Add Go CLAD emitter

This commit adds an implementation of a CLAD emitter for the Go
language.

* Step 2: Add tests for Go emitter

This commit adds tests for the Go emitter. There are two kinds of
tests:
- the stuff in `clad_test.go` was copied from the C# emitter and
  rewritten in Go
- the stuff in `interop.go` (and `cbridge.h/cpp`) tests that Go
  CLAD structures serialized and sent to the C++ emitter, then
  re-serialized and sent back to Go, come back looking the same
  as when we sent them.

The building of these tests rely on a custom version of the .clad
files that live in the DirTreeTests folder, but to make this
commit easier for review, those files are included in the next
commit whose diff can be safely ignored.

* Step 3: Add test CLAD files

* Step 4: Copy new CLAD emitter to victor-clad

* Step 5: Add Go CLAD processing to build

This commit adds the building of CLAD source files located in
`clad/src/clad/cloud`. It also adds the ability for the
`generate_clad` function to specify a target similar to other
versions of the function, and adds generated CLAD files as a
dependency of Go targets.

* Step 6: Add support for multiple dirs in GOPATH

Since CLAD generation now means that Go files can live in multiple
different locations, we need to set our `GOPATH` variable to all
potential source folders so that the Go build can find all
dependencies. This also means fixing some parts of the build that
assumed the `GOPATH` would only contain one directory.

* Step 7: Update mic system to talk to cloud via CLAD messages

Instead of sending particular strings or byte sequences to
communicate with the cloud process, the mic system will now use
structured CLAD messages.

* Step 8: Update AI to talk to cloud process via CLAD messages

User intents will now be sent from the cloud process to the AI
through CLAD messages. Intent parameters are still sent through a
JSON string (since they are not structured data), but the intent
name is now a string.

* Step 9: Update cloud process code to use CLAD

This updates the cloud process code to use CLAD messages to
communicate with outside processes instead of the previous
haphazard protocol of arbitrary byte sequences and strings.

* Add dummy package for cloud clad files to fix `go get`

When building in a clean repo, the CLAD-generated files that the
cloud process depends on doesn't exist yet, so `go get` (the tool
that pulls our Go dependencies) thinks it needs to download them,
then barfs because `clad/cloud` is not a URL. This change adds a
dummy package with no code that will be included in the environment
used by `go get` so that it won't think it needs to grab that
package.