VIC-1703 Track cube with eyes (#1565)
Behavior to procedurally track a cube with eyes horizontally and with
the head vertically. (If we keep this as procedural instead of animated
we'll probably want to move the eyes vertically too.)
Enables use of SetMoveEyes() for ITrackActions, keeping basic actions
the same. I only needed the change for TrackObjectAction, but decided to
keep it at the ITrackAction level to avoid further confusion about when
SetMoveEyes does something. The only usage of it was for laser tracking
and for two SDK messages, the latter of which I removed here.