VIC-2737 Fix Wifi Connection Timer to Reset after Each Connection (#1584)
There was a bug in the wifi connection timeout timer where this timer would work only once per BLE connection because it didn't reset itself correctly. The resulting bug was that in subsequent wifi-connect attempts, Switchboard would send a "Unknown connection status" response back to BLE central immediately after trying to connect (without respecting the requested timeout period). This is not a fix for connection issues, but it will make some of them easier to debug.