Make Hidden Networks Great Again (#1605)
Problem: Hidden networks didn't ever seem to connect, and if they did, it would take like 10 minutes. It seemed that registering the agent to provide SSID/PW wasn't ever actually working, and rather it was only auto connect that was resulting in a successful connection.

Solution: Replace agent dbus object path dashes with underscores. Yep.
Each element must only contain the ASCII characters "[A-Z][a-z][0-9]_"
dbus specification documentation.

With this change, connecting to hidden network seems to work immediately, and the logs are reflecting that the agent is working successfully.