VIC-2800 Behavior for planner smoke tests (#1632)
A behavior for testing basic planner/PathComponent/DriveToPoseAction
functionality.
When the robot sees at least one cube, it constructs poses offset from
the cubes' markers, unless the cube lights are on the side, in which
case it only uses a pose next to that marker. It will drive to that pose
by the cube, then to a pose far from the cube, and back again, forever.
The first stage is to test multi-goal planning in the vicinity of
cubes. Then, when it's driving far away, you can try placing new
obstacles in  front of it, or simply direct the lights side of the cube
toward your undiscovered obstacles. Pick up the robot to reset it.