VIC-2427 Disallow multiple responses from cloud context (#1666)
This bug appears to be a timing issue that happens if the routine
that sends audio is triggered at roughly the same time we receive
an intent back from the server. If the intent is received on one
thread while we're still trying to send audio on the other, that
can cause the audio send to trigger an EOF error (because the
gRPC connection has been closed), which we try to send back to the
cloud process. The process only expects to receive one response
(either an error or an intent) for any cloud context, so this
change will use `sync.Once` to ensure that a response is only sent
one time.