VIC-2973 Don't plan through charger (#1748)
A temporary solution to planning through the charger. This adds the
convex hull of the charger as an obstacle for the planner.
Ideally, the charger is just another ObstacleObservable (VIC-2323).
If we decide not to do VIC-2323, then a new MemoryMapData subclass
should be made so that the charger bounding poly is available, without
having to generate a convex hull. I'm not doing that here.