Clear prox obstacles before returning to charger (#1744)
Hitting the nuclear button. Stale prox obstacles can cause the robot to become stuck while trying to go home. Even with the clearing of prox obstacles that we already do, other obstacles can remain that effectively 'pin' the robot in place, or the robot's estimate of the charger pose is so inaccurate that it does not clear the correct area.