VIC-3019 osState_vicos: Handle no IPv4 address
If the user isn't connected to WiFi, we need to gracefully handle the
case of there not being an IPv4 address.  This change uses `getifaddrs`
to search for an IPv4 address for the `wlan0` interface.  However, if it
isn't found, it simply returns an empty string.  Currently, Victor will
display this as "XXX.XXX.XXX.XXX".