VIC-3019 osState_vicos: Fix leak when getting SSID
The previous code malloc'd memory that was never freed and also was
never used.  This change to the code just gets the SSID of the WiFi
access point we are attached to or returns an empty string if we are
not attached.

(cherry picked from commit e65d1dadb982462fb1082be048c3826778e4438e)