VIC-3097 Fix build errors when building hdrhistogram_c
The hdrhistogram_c library overrides the passed in cflags which are
required to cross compile correctly. This patch appends the original
values in place of overwritting.

Our yocto environment does not create symlinks to SO versioned files yet.
This requires the library version to match the SO version in order for
the dynamic linker to find the library correctly.

Victor runs the install step which causes issues for packages that have
defined install targets. This patch removes install targets from
hdrhistogram_c