Add additional napping transitions (#1775)
Sentiment was that it stays staring at things (in Observing) too long
when nothing is happening, and should nap when in low stim. (The other fix
will possibly be exploring without a known charger.)
This adds transitions into and out of Napping that get us a little closer
toward:

Only a few inputs will get him out of a Sleep state: Wake word, Loud noise,
Being touched, Being dropped/shaken
he needs to seem alive and responsive at least part of the time

Adds transitions to Napping (off charger) at low stim, using only stim
and not time-since-last-face, etc. Also replaces NappingOnCharger
transition conditions to solely rely on stim. This is sort of a test
to see if stim alone is powerful enough to fully control this.
(If we need to bring timer and faces back, there's now a
ConditionFaceKnown.)
Adds transitions out of natural (non-VC) napping if stim rises, or if there's any
type of stimulus.
Creates a ConditionUserIntentActive to conditionally transition if
sleeping started with a user intent
Creates a ConditionAnyStimuli with a standardized set of params that
signify "any stimuli" of particular types.
I also made a ConditionFaceKnown to match ConditionObjectKnown, thinking
I would use that as an exit condition, ended up not using it. Seems
useful so I didn't remove it. Unit tests to come