VIC-3168 Eye animation in anim_weather_windy_01 crashes the robot (#1865)
If we use the NEON box filter in Anki::Vision::Image::BoxFilter() for antialiasing, that causes a crash if an eye intersects with the boundary (that problem is documented in jira VIC-3221). As a workaround, we will now fall back on cv::boxFilter() in those cases to avoid that crash.