VIC-3206 Fix Animation Streamer Bug
In recent AnimationStreamer updates we have added more
reliance on a keyframe’s duration to notify when “complex”
keyframes are done. In the process we added a method that
would set the duration to the beginning of the next keyframe if
it was equal to 0. The fix was to only set the derivations on
specific keyframe types and fix the IsDoneHelper() method to
allow keyframes with the duration equal to 0.

- Fix keyframes with 0 duration being skipped
- Fix keyframe types that don’t need to have duration
- Fix audio debug logs