VIC-3093 Meet victor: Don't turn after physically moved (#1872)
This stops the robot from turning for a configurable amount of time
after it is picked up or rotated when looking for faces during
meet victor.
Changes to the MovementComponent are to know if the robot is being
manually rotated when the wheels aren't moving. This is not the normal
use case of the unexpected movement detector, which checks for
something that prevents intended movement, so it must be explicitly
enabled.