VIC-3226 Fix cross-platform go-grpc/protobuf builds (#1892)
* VIC-3226 Cleanup protobuf (protoc) dependency

Source of the protobuf archives is:
https://github.com/google/protobuf/releases/download/v3.5.1/protoc-3.5.1-linux-x86_64.zip
https://github.com/google/protobuf/releases/download/v3.5.1/protoc-3.5.1-osx-x86_64.zip

These were re-packaged as tar.gz archives because python's zipfile lib does
not preserve permissions on files...

The goal of this change is to handle the protobuf dependency in a platform-agnostic
way.  This will replace the existing use of homebrew on mac and address the fact
that the dependency was completely missing from linux.

CMake files now point to a protobuf version installed under ~/.anki both as a source
of `protoc` and include files required for code generation. This should make builds on
both linux and mac work consistently.

* VIC-3226 Ensure go-protoc exes are built for HOST arch

Go protobuf / grpc generation requires installing (and thus building) a binary that runs on the
host. Since we currently run the install command in the context of the target build, we need to
explicitly pass CC/CXX to tell go to use the HOST compilers. This fixes the linux build in the
yocto environment, where default paths are changed such that "gcc" resolves to an arm
cross-compiler.