VIC-3194 TextToSpeechCoordinator (#1912)
These changes create an engine component to assume the timing and
intent management functionality currently in the SayTextAction in
order to centralize TTS creation and management so that it can be
used directly by behaviors. This enables behaviors to be directly
reactive to TTS generation and playtime, both of which may vary
significantly in length.

The SayTextAction is left intact for now to maintain backward
compatibility with existing features, but it should, at the least, be
stripped out and made to use the TextToSpeechCoordinator interface
internally in the future, if it is not removed.

The TextToSpeechLoop behavior exercises the new coordinator
interface to coordinate animation and TTS activities for a general
procedural TTS case.