VIC-3039: ReactToCliff happening on pickup (#1926)
Cliff-based detection conditions were sometimes not triggering because they were only looking for deltas against a baseline value that could sometimes be acquired while already in the air. In this case, you could actually be holding a robot completely horizontally and suspended in the air without it knowing it was picked up. Simplified the logic to just look for normal cliff detections and if at least 3 are firing then it is considered picked up, regardless of whether or not the wheels are moving.