Fix -I build errors when CONFIGURE=1 is set (#1941)
If `-f` and `-I` were used in conjunction, the build script would
delete generated cmake files (because of `-f`) but fail to re-
generate needed config files (because of honoring `-I`). People
generally weren't using `-f` and `-I` together, but the tricky part
was that other options (like `-D`, which is now used by every
`victor_build_release` build) will have the same effect as `-f` and
force CONFIGURE=1 in the build script.

This change will re-generate the needed config files if CONFIGURE=1
is set.