VIC-3317 Switch ANKI_WEBSERVER_ENABLED to ANKI_NO_WEBSERVER_ENABLED (#1952)
Reverse the logic by request.

Webserver remains enabled by default in a debug build, in a release
build if ANKI_NO_WEBSERVER_ENABLED is not defined then the webserver
is not disabled in code. In a shipping build the webserver is simply not
enabled via the command-line as ANKI_NO_WEBSERVER_ENABLED is defined.

For changes to take affect everyone needs to update their useful ALiases.