VIC-1868 Send status to app with new outbound message types (#1973)
Also VIC-1423, VIC-2924
Proto has messages grouped in a hierarchy so that they can open
streams between the app and robot. Since we're sticking to CLAD on
the robot side, outbound messages that are not part of a
request/response pair need to be grouped as
MessageEngineToGame{ Event{ MessageName{} }}
This moves a handful of meet victor messages over to that format (only
some since someone else is working on them) and creates FeatureStatus
and sub-status messages. These again would need to be written as
MessageEngineToGame{ Event{ Status{ MessageName{} }}}.
To avoid making engine devs care about how GRPC streams are opened
and whether or not something should appear as an app status, there's a
new template helper to automatically construct a MessageEngineToGame
in the above hierarchy directly from a MessageName.
Status, in particular, is broken up into FeatureStatus (the active
feature, also sent to DAS) and a list of other messages that should
appear in the status log. Engine devs shouldn't care about what appears
in the status log, so they can send these messages using the wrapper,
and anyone can later decide to include something in the status log by
modifying the Status union in clad.
This also adds a RobotHistoryRequest and RobotHistoryResponse for the
app to get recent status messages. This is not wrapped in Event, since
it is a request/response pair, and the templates figure that out
for you.