VIC-3417 Dance to the beat: Listen for beats during "eyes only" anims (#1984)
Previously, when DTTB behavior started, we completely stopped listening for beats and just extrapolated forward in time to estimate the times of future beats. This was because the 'beat' animations typically move a bunch of motors and ruin the microphone input. However, certain animations (e.g. the "eye beat" animations) either do not move the motors very much or at all, and so we should be able to listen for beats during that time.

- Allow active listening for beats during dance phrases marked `canListenForBeats`.
- Add ability to register/de-register `OnBeat` callbacks within `BeatDetectorComponent`.
- Disable the get-in animation since it may cause the motors to move too much to continue to accurately detect a beat.
- Increase the number of "eye beat" animations and allow listening for beats while they are playing.