VIC-2991 Add gRPC messaging for SDK to request control from behavior system #2003
Add gRPC messages for the SDK layer to request that the SDK behavior be activated.

This commit gets the gRPC and clad messages in place. The gameToEngine messages are not yet used by the engine, and the engineToGame messages are not yet sent by the engine.

Note than in the behavior system, an activation request doesn't fail; it just waits until it gets activated. Later I will be adding a configurable timeout on the Python side and C# side so that if the SDK behavior doesn't get activated within the default time period, then the the SDK will cancel the activation request.