VIC-3345 VIC-3397 VIC-3399 Onboarding framework (#2017)
Initial behavior has changed:
A new BehaviorsBootLoader component determines how the behavior stack
is initialized when it boots. Robots now boot into either a normal
stack, an onboarding stack, a factory stack, or, for those attached to
shaking the robot, DevBaseBehavior. Normal and onboarding stacks both
start with the same sleeping eyes animation, which may need to change
to match the proposed "closed eyes" fixed image to be displayed while
the OS boots. Except for factory, the stack that is chosen is based on
the current onboarding state, which includes a special value for
DevBaseBehavior. Bottom line is that everyone's robot will boot into
a pre-onboarding stage by default, where the robot doesn't move, but
where its eyes are just slivers. If you want to see eyes, you can use
a console var to set the onboarding state to Complete or DevDoNothing,
depending on whether you want to shake the robot. It will be saved to
persistant storage.

Onboarding framework:
Creates a framework for onboarding progression in which behaviors are
grouped into stages. A stage progresses linearly based on app and
robot input, and it also defines a checkpoint in case of crashes or
reboots. An IOnboardingStage handles simple logic of what behavior
instance it wants to run based things like on app input (e.g.,
"Continue"), dev tools input, or when another behavior ends.
The coordinator behavior BehaviorOnboarding manages these stages,
interruptions shared among stages, messages with the app and devtools,
and behavior delegation. This framework uses explicit logic rather
than latent transitions among behaviors to ensure a precise experience.

VIC-3345 Stage logic and stub behaviors for the wake up stage.

VIC-3397 Stage logic to progress through existing Meet victor behaviors

VIC-3399 Console vars for QA to send all of the imperative message
types we expect the app to send the robot during onboarding, like
"Continue" and "Skip". There's a table of what button should do what
in the ticket.