VIC-3078 Partial fix for not turning after voice command (#2117)
The dominant mic direction is the direction with the highest count.
Whenever the robot is moving, we accumulate the "unknown" direction,
which often dominated the others.
The forward direction still often dominates because of audio from the
speaker. But, if the robot isn't playing a ton of audio, this is an
improvement.