VIC-1394 Make victor_build_* aliases call scripts
For the Victor OS build of victor, we want to build with the same
build flags that developers are using when they build for shipping,
release, etc.  Unfortunately, aliases cannot be invoked in bitbake
shell scripts (or any shell scripts for that matter).  To fix this, I
have created new victor_build_*.sh build scripts that contain the
build flags.  The victor_build_* aliases now just invoke the scripts.