SAIP-1523 Hard fail w/ code if robot fails certificate check (#2116)
If a robot fails verification of its factory certificate data,
the robot should hard fail in a way that will be obvious to people
around the office so we can become aware of any issues. With the
declining prevalence of DVT2 robots (and the fact that certs can
be installed on older robots anyway if they still need to be used),
it's finally time to make this change.