VIC-3710: Add wifi scan check to playpen (#2134)
This change adds messaging to allow the engine to request a WiFi scan
from switchboard. This is intended to be used only as a fallback test
of wireless capabilities in the case where the BLE cube test causes
Playpen to fail erroneously.  Because of the limited use case, switchboard
only returns the count of ssids present in the scan results.

Test should be disabled in pre-playpen fixture by writing to WifiScanError bit in EMR.playpenDisableMask